Commit built files (#52)
It has been very onerous to try to use the Middleman build tools, as our configuration code relies on versions that are very out of date (i.e. Ruby 2.1.5). We have relatively complicated Docker setup to try to deal with that, but even that is now difficult to use. There's no telling when or if we will need this repository to be maintainable, so we can just make changes directly to the flat files. Going forward, all changes should be committed directly in /build.